cli: allow use of non-fully-qualified job id's
job {list,status,deploy,undeploy,remove} can now
accept job id's in the form of just the job name
or the job name and version. The job is first
resolved using a /jobs?q=... query and depending
on the type of the command an action is taken on
the resulting job id or aborted if the job id is
ambiguous and returned many jobs.