port mapping
Static and dynamic port mappings can be defined for a job.

For dynamic port mappings, the agent will automatically allocate an
available port on the host. Note that this implies that the selected
ports for a job will vary between hosts. The allocated port can be
read out from the task status using the agent or the cli.

For static port mappings, the agent will always instruct docker to
route the specified public port on the host to the specified private
port in the container.