Don't print stack trace by default
When exceptions propagate up to CliMain, only the exception message
will be printed. You can print the entire stack trace with -v.

The exceptions thrown when you can't connect to the master are now
handled by ControlCommand so that we can print out which master
or site we were trying to connect to.

A message like this will be printed if a specified was site:
"Request timed out to master in ash.spotify.net (srv://helios.services.ash.spotify.net)"

A message like this will be printed if an endpoint was specified:
"Request timed out to master tcp://master.ash.spotify.net:5800"

The version check in client is now asynchronous. This allows
RequestTimeoutException and SendFailureException to be caught and
handled by ControlCommand, without having to add 'throws' clauses to
the client API since they are checked exceptions.

Change-Id: I25062d9b175d888ff7330791b28ed27cef6bfd8c