Don't catch InterruptedException in ControlCommand
If run would throw an InterruptedException, ControlCommand would
catch it and continue trying other targets. Until we have figured
out how to handle multiple targets when there is a failure, it is
safer to simply quit. This also makes sense since we do not catch
and continue when an IOException is thrown.