agent -> host
Refer to hosts running jobs as hosts instead of agents. Let the agent
refer to the service that controls the host(s).

This pushes the implementation detail that agents run on every host
well into the agent itself. The master can now almost completely
work only with jobs and hosts.