Refactoring of SystemTestBase to facilitate integration tests
SystemTestBase has been a bit of a dumping ground of stuff.  This
doesn't fix that entirely, but does tidy things up a bit and I've
at least verified that all fields at least are used, and have
tightened up the access qualifiers so when reading it, you can have
a better idea of what's used where.  A few fields that were accessible
have been made private and have had accessors made for them so that
when running as integration tests, they provide something different
than they used to.

All the new integration tests consist of empty classes as SystemTestBase
figures out what's going on and does the right thing.

To run them, set the HELIOS_ENDPOINT environment variable.  If you want
it to not clean up it's jobs when it's done, set the ITCASE_PRESERVE_JOBS
environment variable.  A sample commandline would be:

    HELIOS_ENDPOINT=tcp://localhost:5801 mvn -P integration-only verify

This makes it so it *only* runs the integration tests, and not the unit
tests.