zk retries with uuid's
ZK connection is often spuriously lost, causing
helios operations to fail. Mitigate this by having
curator do retries. In order to make the reason
for conflicts of create operations unambigous we
use creation operation nodes with uuid's for jobs
and deployments. When a conflict exception is
caught (typically NodeExists) for e.g a job
creation transaction we can check for the
existence of the uuid'd creation node to know if
the conflict was due to a retry or a real conflict
with a concurrent job creation.

Thus we'll never get spurious "Job already exists"
errors in the cli for job create commands.