Fixed JobHistoryTest flakiness - due to race in Supervisor
This time it's due to a race in the posting of the task status
within the Supervisor/TaskRunner.  Depending on ordering, the
Stop command may finish before the container exits and the
TaskRunner posting EXITED, but not always.  Probably not worth
it to create a defined ordering for this, so having the test
accept either seems best for now.