Run things as the newly created helios user
Have to, at least temporarily, hard code in the homedir here in the postinst
scripts as /spotify/var/helios, so that ownership of the directory is correct
across the conversion and helios restart. And to not have to put some kind of
complicated temporary thing in puppet.

Also, we should consider setting the default data dir for helios to some
standard place (i.e. not /spotify/...) for going opensource anyhow.