agent: reap unexecuted tombstoned tasks
The agent would not properly remove tombstones,
i.e. tasks with UNDEPLOY goals, if they had not
been entered in the set of executions. Deploying
and immediately undeploying a job on a host
before the agent picks up on the START goal would
then cause the tombstone to never be removed.