Respawn forever, but only once every 5 seconds
Upstart's default behavior is to stop trying to respawn a job if it fails
too often, but we want to keep trying to respawn forever so that if the
underlying error condition is fixed, Helios comes back up automatically.

That said, we throttle respawns to no more than once every 5 seconds via a
post-stop script to keep things from being too crazy.