Added TemporaryJob.address helper method
Before we only had an addresses method which returned a list of hosts
the job was deployed to, but most of the time users will only deploy
to a single host. Instead of making them pick the first element from
the list, we added an address method which returns the first item in
the list. If there is more than one item an AssertionError is thrown.