TemporaryJob: enable deploy without specifying a host
Instead of explicitly specifying a host to deploy to, calling deploy() with
no arguments picks a random host from the Helios cluster to deploy to.

We require specifying a host filter pattern, either programmatically or in
the HELIOS_HOST_FILTER environment variable. Only hosts from the Helios
cluster whose names match the filter pattern will be considered when
picking a host.