HELIOS_HOST_FILTER was being ignored by TemporaryJobs
When choosing which host(s) to deploy to, TemporaryJobs
would use HELIOS_HOST_FILTER to come up with a list of
hosts matching the filter. But when it came time to deploy,
it would ignore this list, and use the list of all hosts
instead, effectively deploying to a random host.