[helios-tools] make it so deploy/undeploy can take non-fqdn names
Specifically, if you specify a unique prefix, it will [un]deploy to
the matching host.  If you specify a non-unique prefix, it will
look at the matches and the one whose domain appears first in the
dns search path will be returned.  This works much more intuitively
than I'm explaining.