Tagged commits create Github Releases
Any tagged commit that passes tests will create a GH release and
upload a tarball with the Debian packages created.

Branch restrictions don’t work due to a bug with Travis although at the
time I don’t think this will be too much of an issue, we don’t tag that
much, only for releases.