HeliosClient should return JobDeleteResponse, not null, when deleting nonexistent job
Currently if you try to delete a nonexistent job, HeliosClient returns null,
so the caller doesn't know what happened. This caused a NPE in TemporaryJobs,
because it tried to delete a nonexistent job, and didn't expect the client to
return null. It now returns JobDeleteResponse with the status set to
JOB_NOT_FOUND so the caller can decide if this should be handled as an error
or not.