Workaround for btrfs issue with removing containers
When using btrfs in certain environments (like CircleCI), the btrfs driver
throws an exception when attempting to remove a container. However, the
container does actually get removed.

Check for this specific case in our tests. This should only affect
AgentRestartTest, since it is the only thing that actually attempts to
remove a container.