[pom] add two profiles to control signing and test
When testing the release process, or running the release bits (but not
deploying), it can be desirable to disable gpg signing and the running
of tests.

This change also reinstates part of the pom that was erroneously removed
in 5443b556d9ad6621815862861078875d2264402d, which turned off integration
testing when performing the release.