Improve log messages from helios testing framework
Many of the log messages produced by the testing framework include
the helios job ID. But these IDs are auto-generated and very long,
so they clutter the logs and don't provide much value to the user.
When the use declares a TemporaryJob, they specify the image they
want to deploy, so it's more useful to log the image name.

Changed many of the detailed log messages from info to debug.

When waiting for an image to start, we would only log if the
image is pulling. We now log whatever state it's in, so we
always know the state of things.