Add more info to testing framework logs
Before, if a container failed to start we would throw an exception
with the error message, but we would say which job threw the
exception. This was confusing to debug when you've deployed multiple
containers. We now throw an exception with this message:

Unexpected job state EXITED for job tmp_5ebc2788_spotify_cassandra:c008a672:bc1ebbf
with image spotify/cassandra on host heliosciagent Check helios agent logs for details.

Also when pulling an image takes a long time your screen would
fill up with the message "Job state: PULLING_IMAGE". When pulling
and image takes a long time, it would be easy to lose context and
know which image is being pulled. The message is now this:
Job state of spotify/cassandra: PULLING_IMAGE