Removing unused dependencies
Found a couple dependenices which aren't being used. The jsr305 dependency
in helios-testing was causing a problem for one of our users because it
conflicts with a lower version they are using. Since we're not using it,
might as well remove it and not force them to exclude it in their pom.