Replace ${revision} with 0-SNAPSHOT in pom files
The point of using ${revsion} was so that the build system could replace
it at runtime with the build number by passing -Drevision=BUILD_NUMBER.
This works while maven is building your code, but causes issues when doing
a maven install. This is because it doesn't actually substitute the value
on disk, and maven install copies the actual files. So you end up with jars
which are dependent on com.spotify:helios-parent:pom:0.8.${revision}, which
causes a failure when the jar is used.

To get around this, our build machine already replaces ${revision} with the
actual build number before invoking maven. So let's just replace ${revision}
with 0-SNAPSHOT so local installs work too. The build machine can just search
and replace this new string instead.