sort JobListCommand output in non-quiet mode also
JobListCommand currently has logic for sorting the output by the jobId
when the `quiet` flag is set to true. For regular output, the output is
unsorted (the actual order is determined by iteration order of the
HashMap used to store the status futures), but it would be just as
useful to have the output ordered by jobId in this command mode also.