Add support for multiple container decorators
TaskConfig now accepts a list of container decorators, which are applied in
order when creating or running a container.

Previously, we only supported having a single container decorator for a
Helios job. This worked out since we only have one real decorator, the
SyslogRedirectContainerDecorator, but is a problem for upcoming features.