solo: Inject right environment in integration test
Instead of looking at `DOCKER_HOST` and stuff from the environment of the
test runner, we need to be getting it from the environment of the host on
which the helios-solo job is run.

As part of this, we now have Helios agents publish the actual Docker host
and Docker certificate path they're using.

This fixes test failures when running `HeliosSoloIT` against a remote
Helios cluster (as opposed to a local helios-solo instance).