More fixes for the HeliosSoloIT
* If the DOCKER_HOST is 127.0.0.1, fall back to Unix sockets instead.

* Use the latest version of docker-client, which supports checking the OS
  name directly rather than kernel version (for Boot2Docker detection).

Note that as part of the docker-client upgrade, we need to declare an
explicit dependency on commons-lang:commons-lang:2.6 to resolve a conflict
between docker-client and another package.

Also had to remove helios-testing's dependency on com.github.jnr:jnr-posix.
I have no idea why the helios-testing project was depending on a Unix
socket library, it really shouldn't be and was causing a conflict.