Restart container if it dies during healthchecking
If a container dies during healthchecking — or at any other time — we
should restart it. Due to a bug, we were not restarting containers that
died while in the HEALTHCHECKING state.

Now, as part of the health check, we ensure the container is still running.
If not, we throw an exception and go to FAILED, which eventually causes the
container to be restarted. Fixes #440.

This also includes a fix for flakiness in `HealthCheckTest`, wherein the
test container would exit immediately after passing the health check. This
could cause the test to fail. Instead, we now make sure the test container
stays up and listening.