Add a managed deployment service
This adds a new first-class concept to Helios called deployment-groups,
which form the basis for doing rollouts that are orchestrated by Helios.
A deployment-group is defined by a set of labels that determine what
agents are part of the deployment-group, and a job which should be
deployed on those hosts.

The target state (job) for the deployment group is set via the
rolling-update CLI command. At any time a deployment-group will have
at most one job* deployed to a host in the deployment-group.

*: Transient exemptions to this statement may apply!

To give an example:

  helios create-deployment-group my_group foo=bar

Creates a deployment group, comprising all agents with that are marked
with the foo=bar label. At this point the deployment-group will not
deploy any jobs to any of the hosts. Running

  helios rolling-update my_job:1 my_group

will deploy the "my_job:1" to all agents in the deployment-group, one
at the time (by default). Running rolling-update a second time with
a different job will cause the previous version of the job to be
undeployed and replaced by the new job. E.g.

  helios rolling-update my_job:2 my_group

will cause the my_job:1 job to be replaced with my_job:" on all the
agents in the deployment-group.