Maintain a list of hosts for each deployment group
Previously, the list of hosts for a deployment group was inferred based on
the rollout tasks for the group.

But this isn't how things work conceptually. A deployment group is defined
by a set of labels, and the hosts that are in it are based on those labels.

Therefore, we now have a separate reactor that determines the hosts for all
deployment groups and writes that to ZooKeeper. Then, for a rolling update,
we can just look at that list of hosts in ZooKeeper instead of having to
calculate it on the fly.

This also has the advantage of allowing the CLI to report all hosts that
are targeted by a deployment group, even if a rolling update hasn't been
triggered or if the rollout plan hasn't been calculated.