Kill container based on name if ID not available
If a job is undeploy right after it's deployed, we may have asked Docker to
create a container but still not gotten the ID back. In the past, we would
then leave this orphaned container around.

Now, if we don't have a container ID, we attempt to kill any container that
matches the predetermined container name to avoid this race condition.