HealthCheckTest: Fix intermittent failures
The HealthCheckTest runs netcat in some containers via TemporaryJobs and
then probes those netcat instances to make sure they're reachable. When
netcat is probed, it usually exits immediately. Generally we wouldn't care
sicne the test ends right after the probe.

However, if TemporaryJobs noticed that the job exited while the test was
still running, it would throw an exception due to the job being unhealthy.
The test would then fail.

So now, we run netcat with `-k` so it sticks around as a persistent server
and the container doesn't exit prematurely.

Fixes #529.