Make QueueingHistoryWriter work for generic events
The QueueingHistoryWriter was used to write task status events to ZooKeeper
and Kafka. Unfortunately it was very coupled to TaskStatusEvent.

Now we've moved it to servicescommon and made it generic and abstract so it
can be used for writing all types of events. The class that writes events
for tasks is now named TaskHistoryWriter and extends QueueingHistoryWriter.