Fix mocking in TaskHistoryWriterTest
It turns out you're not supposed to change the stubbing on Mockito mocks
once you start calling the mock from different threads. Doing so is unsafe
and can cause failures.

We were changing the stubbing on the mocked ZK client after using it in
TaskHistoryWriter, which caused weird exceptions and test failures. Fix
this so we do all the stubbing beforehand.