Move CLI list master logic to server
This commit moves the logic for listing helios masters from the CLI to
the server (see the much shorter `MasterListCommand.java` class).

JSON is emitted by calling a different HTTP endpoint and just having the
client print out the string it receives in the HTTP reply.

Benefits are a simpler client, less chance of client-server
incompatibilities, and potentially better performance due to masters'
proximity to the ZooKeeper cluster.

If we like this idea, we can do this for the other CLI commands.