add a 'metadata' field to Jobs
The "metadata" field allows users to optionally store an arbitrary list
of key-value pairs to be associated with the Job.

The Helios master or agent doesn't act on the contents of the Job
metadata, but the field can be used for storing information alongside
the Job definition such as Git Commit ID, etc.