cosmetic changes to user manual / All Job Options
reformatted the `All Helios Job Options` section of the manual to change
the markup from a list of bullet points to a sub-heading for each job
field. The intent here is to make the text on the page less dense, for
example when trying to view/understand the format of the port mapping or
service registration sections (including the many examples and defaults)
which is rendered as a big blob of text in HTML.

A nice benefit of having headers for each field is that it becomes
possible to link to each section on Github as well.