additional fixes for job metadata / GIT_COMMIT
This fixes failures in the system tests when run under Jenkins (or any
agent with the `GIT_COMMIT` environment variable set) that are caused by
the job being created with the environment variable automatically
included but the test does a deep-equality assertion against an expected
Job value.

The hacky fix here is to modify `assertJobEquals()` to remove any
automatically-included-metadata from being compared in the assertion.