refactor CliConfig to remove static reference to System.getenv()
Storing the result of `System.getenv()` in a non-private static
reference causes the tests that touch `CliConfig` to need to worry about
setting or resetting this reference for the environment variables it
wants to simulate as being present for it's test case.

Refactoring `CliConfig.fromUserConfig()` to require the caller to pass
along what it thinks it's environment variables are (which is only
called from one place in non-test code) removes this need for the tests
worry about cleanup and state.