Fix flaky DeregisterTest
* Fix NPE in AgentZooKeeperRegistrar.tryToRegister() that caused this
thread to die silently. This caused a race which sometimes failed the
tests.

* Wait for first host in
DeregisterTest.testRegistrationResolutionTtlNotExpired() and
DeregisterTest.testRegistrationResolution() to come up AND report
HostInfo to ZK in order for second host to detect when the
first host reported its HostInfo back to ZK so the second can properly
deregister the first.