servlet filter for checking access tokens
When the master is started with `--auth-required-versions`, `AccessTokenFilter` checks each HTTP request to the master for the presence of an access token in the http headers.

If the http header is not present, a response of 401 Unauthorized is
sent.

If the http header is present, the supplied access token is checked
against the `HttpAuthenticator` that the master is configured to use. If
the authenticator indicates that the token is not valid (or has
expired), a response of 401 Unauthorized is sent.

Otherwise the request is allowed to continue.

Adding this filter requires a few changes in the HttpAuthenticator
interface:

- add a method to expose the name of the auth scheme being used so that
  it can be sent to unauthenticated clients in the WWW-Authenticate
  header
- add a method for validating the supplied access token for the user's
  request

some TODOs:
- Not clear where the `username` to use when validating the token will
  come from in the http request
- Should probably link the new argument controlling whether
  authentication is enabled with the loading of a plugin that actually
  does the authenication - enabling authentication but using the no-op
  implementation seems pointless
- is the signature for `verifyToken(username, token)` legit for the way
  that crtauth works?