ability to load AuthenticationPlugin from any path
Implements the `--auth-plugin` part of the Helios master where an
AuthenticationPlugin instance can be loaded from an arbitrary path, i.e.
from a user-supplied JAR file not packaged with Helios itself.

To test this, also added an integration test harness using
maven-invoker-plugin which first builds the JAR containing the plugin to
be loaded and then executes a test of loading that JAR.

The important part of the test, and the reason why this is done with
maven-invoker-plugin, is that the plugin being loaded should not be on
the JVM's actual classpath, but only accessible through the `pluginPath`
configuration. It's not trivial or obvious how to do this in a
traditional Maven build (where you would build a separate JAR, then run
some tests *without* that JAR on the CLASSPATH); the
maven-invoker-plugin comes in handy for this sort of behavior.

Thanks to @dflemstr for adding similar behavior to missinglink which is
where the inspiration for this test came from:
https://github.com/spotify/missinglink/pull/32