SystemTestBase: don't execute CLI commands to ensure master is up
Executing the CLI command breaks if we want to have integration tests
that start a Master with authentication enabled.

Changed the method to instead test the service's healthcheck endpoint
and renamed it to make it clear that it is actually making sure that the
master is fully up (rather than just the zookeeper connection part).