build api-docs module after helios-testing
The Jacoco setup in the api-docs module requires helios-testing to have
been build first (as it refers to it's `target/classes` directory).

Adding a `<dependency>` on helios-testing moves the API Documentation to
the last module to be built by Maven.

But in reality we are probably abusing Maven and the setup of this thing
is totally fragile.