allow system-tests to set custom environment variables
Refactor MasterService so that it is possible to pass custom environment
variables (represented as a `Map<String, String>`).

This makes it possible for system tests to start the master with custom
defined environment - previously not possible since the MasterMain is
started in the same process.

Also pass the Master's view of environment variables down to the
AuthenticationPlugin framework to simplify their loading logic.