test that users can proactively authenticate
Adds a system test that enables authentication with
`--auth-minimum-version` set to a really high version (99.9.9) and
verifies that a client can supply `Authorization` headers and be
authenticated if they really want to.

To verify this behavior, add a Jersey response filter that sets a header
`Authentication-Principal` containing the username of the authenticated
user.