allow for eager client authentication
Change the client authentication behavior so that the flag
`--force-authentication-scheme` will cause the client to immediately
authenticate using the specified scheme.

If not set, then the previous behavior of waiting until the master
returns a `401 Unauthorized` response is kept.

To enable this, a small refactoring to the `AuthProvider.Factory`
interface: change the first parameter of the `create` method to `scheme`
instead of `authorization header value` - this moves the header parsing
up one layer.