client: Use ssh-agent identity for TLS client auth
Use the local SSH agent identity to generate a self-signed certificate for
TLS client authentication, and sign the handshake hash with ssh-agent.

This lets the client setup a secure SSL connection and prove ownership of
its SSH public key. It is up to the server to verify the public key against
a known-good list (like in LDAP).