Allow specifying a job JSON for a temporary job
Currently folks store their job configurations for their real jobs in JSON
files, but configure their temporary testing jobs programmatically. This
has the downside of making the job that's run in integration tests
different from the job that actually gets deployed.

To address this, allow creating a TemporaryJob from a job config JSON.
Ideally this should be the same JSON as the real job. For example:

    temporaryJobs.jobWithConfig(".helios/job_config.json").deploy()