fix type error in KAFKA_QUOROM_PARAMETER
The Kafka `ProducerConfig` expects the value for `ACKS_CONFIG` to be a
`String` and not an `int`, even though the value of the String looks to
almost always be numeric (seems like it is a String since one possible
value is `all` in addition to `1` or `0` or `-1`). This was introduced
in ae50b2f.

With this configuration error, the agent or master is never actually
able to instantiate a KafkaProducer and not able to send any events to
Kafka.