Fix race conditions that might lead to inconsistent DG state
There were two related race conditions that could cause a DG get
inconsistent state, both triggered by removing the DG. The issue
manifests itself with a DG that has no config node but does have
the status and/or tasks paths.

This could happen if a remove happened at the same time as a
rolling-update was issued or if the host-update thread updated
the hosts for that DG as the DG was removed. The latter is much
more likely to happen.