allow SSL hostname verification to be disabled
Adds a CLI switch like curl's `-k`/`--insecure` to disable hostname
verification when using HTTPS to communicate with the Helios master.

This can be useful when attempting to connect to a single master
instance using `-z` where the certificate presented by the master has a
subject that doesn't match it's hostname (for example, if the
certificate's subject is `helios.spotify.net` but the server's hostname
is `server1.foo.spotify.net`).

To make the HostnameVerifier logic testable, extracted it's use in
DefaultRequestDispatcher to it's own class.