RetryingRequestDispatcher: log why the request failed
RetryingRequestDispatcher catches failed futures returned by other
RequestDispatchers and retries the request a certain number of times.
Only after the final failure is any of the exceptions propagated.

To help in debugging, at least log the exception that occurred on each
try (at debug level). This is also useful if the user kills the
client/CLI when the requests are failing rather than waiting for all of
the retries to happen.