Ensure we're recording the TLS handshake correctly
By default, BouncyCastle keeps a rolling hash of TLS handshake messages and
then signs that hash as part of the handshake. Unfortunately, that doesn't
work for us when using ssh-agent, since ssh-agent does the hashing itself.

So instead, we keep a complete record of all TLS handshake messages, and
send that to ssh-agent for hashing and signing.

Since we are keeping this record ourselves, and since only certain types of
messages should be recorded, there's potential for arcane bugs. To prevent
head-scratching, we can compare the SHA1 hash of our record to the rolling
hash that BouncyCastle has kept. The two should match, and if they don't,
then we know our recording is being done incorrectly.