fix infinite loop if masters return 502 Bad Gateway
AuthenticatingHttpConnector handles a 502 Bad Gateway response by just
trying another iteration of the `while (true)`.

If for some reason the endpoints always return 502 (for example, the
service behind nginx is down) then this means the
AuthenticatingHttpConnector will loop forever (as long as the endpoints
are returning 502) with no way to interrupt.

This check for 502 Bad Gateway is better made in the
DefaultHttpConnector as it has nothing to do with authentication. A
response with this status code will have an exception thrown so that it
is caught by RetryingRequestDispatcher (and a new request is scheduled).