fix infinite loop if masters return 502 Bad Gateway
AuthenticatingHttpConnector handles a 502 Bad Gateway response by just
trying another iteration of the `while (true)`.

If for some reason the endpoints always return 502 (for example, the
service behind nginx is down) then this means the
AuthenticatingHttpConnector will loop forever (as long as the endpoints
are returning 502) with no way to interrupt.

This undos a change in 27d3c25 where a 502 caused the http client to
just try again with a new endpoint (and which was carried over to the
HttpConnector stuff). This is not needed anymore since
RetryingRequestDispatcher was introduced in 1c2c451.