address race condition in HeliosSoloDeploymentTest
Expose a port in the test container so that the call to `deploy()` does
not finish until the container is fully up (and the port is successfully
probed).

This is to avoid the rest of the test checking how many jobs
exist before the container itself for the job is fully started.

Changed the image used in the job from busybox to onescience/alpine as
busybox has a weird netcat version that will exit as soon as a
connection is received when `-l` is used.