do not fail on errors setting up client certificate
Refactors the Certificate and PrivateKey setup done in
`CertificateHttpsHandler` so that the two are returned from a single
method call; this makes it easier to catch errors setting up either of
them in a single place.

Adds a boolean switch that controls if the Exceptions from setting up
the client certificate from a file are ignored or not.