Use signed certificates for ssh-agent auth
Previously, we were overriding a bunch of things to let the ssh-agent sign
the TLS handshake on-the-fly. While this works, it's a lot of code and also
required us to duplicate/override parts of the TLS internals.

Instead of doing that, we now just generate a new key pair and certificate
and have the ssh-agent sign that certificate. This is much less code, much
simpler, and is actually compatible with our existing infrastructure.

It also consolidates a lot of the code for the ssh-agent and certificate
file options, since both just generate a java.security.cert.Certificate and
a java.security.PrivateKey which is handed off to the standard
SSLSocketFactory implementation.