log system test results to file in CircleCI
Since the tests are run in parallel in Circle with the surefire setting
`forkCount=4`, it is hard when looking at failures of anything extending
from SystemTestBase to understand which log lines in standard out are
from which test. So log these to a file instead, where there is one file
per test method.

Also: use the documented environment variables `$CIRCLE_ARTIFACTS` and
`$CIRCLE_TEST_REPORTS`.